+++ src/endstone_endweave/codec/__init__.py
@@ -21,6 +21,8 @@ from endstone_endweave.codec.types import (
     UVAR_INT64,
     VAR_INT,
     VAR_INT64,
+    VEC2,
+    VEC3,
     ByteArrayTag,
     ByteTag,
     CompoundTag,
@@ -84,6 +86,8 @@ __all__ = [
     "UVAR_INT64",
     "VAR_INT",
     "VAR_INT64",
+    "VEC2",
+    "VEC3",
     "read_nbt",
     "write_nbt",
 ]